Moving Table to utils. Fixing couple of small things
