+++ util/src/lib.rs
@@ -144,6 +144,7 @@ pub mod log;
 pub mod panics;
 pub mod executor;
 pub mod keys;
+pub mod table;
 
 pub use common::*;
 pub use misc::*;
